+++ src/views/Stake.vue
@@ -1,0 +1,3 @@
+<template>
+  <h1>Stake</h1>
+</template>
